refactor(folder-browser): G.87.3 — extract useFolderBrowserNowPlayingPath hook
